Reorganize version solver code (#1749)
This splits up the monolithic version_solver.dart file into one class
per file, and renames BacktrackingSolver to VersionSolver since
there's only one solver implementation.
