+++ lib/src/global_packages.dart
@@ -23,7 +23,7 @@ import 'package.dart';
 import 'package_name.dart';
 import 'pubspec.dart';
 import 'sdk.dart' as sdk;
-import 'solver/version_solver.dart';
+import 'solver.dart';
 import 'source/cached.dart';
 import 'source/git.dart';
 import 'source/hosted.dart';
