+++ lib/src/solver/unselected_package_queue.dart
@@ -9,7 +9,7 @@ import 'package:stack_trace/stack_trace.dart';
 
 import '../log.dart' as log;
 import '../package_name.dart';
-import 'backtracking_solver.dart';
+import 'version_solver.dart';
 
 /// A priority queue of package references.
 ///
@@ -21,7 +21,7 @@ class UnselectedPackageQueue {
   SplayTreeSet<PackageRef> _set;
 
   /// The version solver.
-  final BacktrackingSolver _solver;
+  final VersionSolver _solver;
 
   /// A cache of the number of versions for each package ref.
   ///
